Move EGL helper functions into their own module
