+++ shell/platform/linux/fl_renderer.cc
@@ -5,6 +5,7 @@
 #include "fl_renderer.h"
 
 #include "flutter/shell/platform/embedder/embedder.h"
+#include "flutter/shell/platform/linux/egl_utils.h"
 
 G_DEFINE_QUARK(fl_renderer_error_quark, fl_renderer_error)
 
@@ -19,284 +20,6 @@ typedef struct {
 
 G_DEFINE_TYPE_WITH_PRIVATE(FlRenderer, fl_renderer, G_TYPE_OBJECT)
 
-// Gets a string representation of the last EGL error.
-static const gchar* get_egl_error() {
-  EGLint error = eglGetError();
-  switch (error) {
-    case EGL_SUCCESS:
-      return "Success";
-    case EGL_NOT_INITIALIZED:
-      return "Not Initialized";
-    case EGL_BAD_ACCESS:
-      return "Bad Access";
-    case EGL_BAD_ALLOC:
-      return "Bad Allocation";
-    case EGL_BAD_ATTRIBUTE:
-      return "Bad Attribute";
-    case EGL_BAD_CONTEXT:
-      return "Bad Context";
-    case EGL_BAD_CONFIG:
-      return "Bad Configuration";
-    case EGL_BAD_CURRENT_SURFACE:
-      return "Bad Current Surface";
-    case EGL_BAD_DISPLAY:
-      return "Bad Display";
-    case EGL_BAD_SURFACE:
-      return "Bad Surface";
-    case EGL_BAD_MATCH:
-      return "Bad Match";
-    case EGL_BAD_PARAMETER:
-      return "Bad Parameter";
-    case EGL_BAD_NATIVE_PIXMAP:
-      return "Bad Native Pixmap";
-    case EGL_BAD_NATIVE_WINDOW:
-      return "Bad Native Window";
-    case EGL_CONTEXT_LOST:
-      return "Context Lost";
-    default:
-      return "Unknown Error";
-  }
-}
-
-// Converts an EGL decimal value to a string.
-static gchar* egl_decimal_to_string(EGLint value) {
-  return g_strdup_printf("%d", value);
-}
-
-// Converts an EGL hexadecimal value to a string.
-static gchar* egl_hexadecimal_to_string(EGLint value) {
-  return g_strdup_printf("0x%x", value);
-}
-
-// Converts an EGL enumerated value to a string.
-static gchar* egl_enum_to_string(EGLint value) {
-  if (value == EGL_FALSE)
-    return g_strdup("EGL_FALSE");
-  else if (value == EGL_LUMINANCE_BUFFER)
-    return g_strdup("EGL_LUMINANCE_BUFFER");
-  else if (value == EGL_NONE)
-    return g_strdup("EGL_NONE");
-  else if (value == EGL_NON_CONFORMANT_CONFIG)
-    return g_strdup("EGL_NON_CONFORMANT_CONFIG");
-  else if (value == EGL_RGB_BUFFER)
-    return g_strdup("EGL_RGB_BUFFER");
-  else if (value == EGL_SLOW_CONFIG)
-    return g_strdup("EGL_SLOW_CONFIG");
-  else if (value == EGL_TRANSPARENT_RGB)
-    return g_strdup("EGL_TRANSPARENT_RGB");
-  else if (value == EGL_TRUE)
-    return g_strdup("EGL_TRUE");
-  else
-    return nullptr;
-}
-
-// Ensures the given bit is not set in a bitfield. Returns TRUE if that bit was
-// cleared.
-static gboolean clear_bit(EGLint* field, EGLint bit) {
-  if ((*field & bit) == 0)
-    return FALSE;
-
-  *field ^= bit;
-  return TRUE;
-}
-
-// Converts an EGL renderable type bitfield to a string.
-static gchar* egl_renderable_type_to_string(EGLint value) {
-  EGLint v = value;
-  g_autoptr(GPtrArray) strings = g_ptr_array_new_with_free_func(g_free);
-  if (clear_bit(&v, EGL_OPENGL_ES_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_OPENGL_ES_BIT"));
-  if (clear_bit(&v, EGL_OPENVG_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_OPENVG_BIT"));
-  if (clear_bit(&v, EGL_OPENGL_ES2_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_OPENGL_ES2_BIT"));
-  if (clear_bit(&v, EGL_OPENGL_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_OPENGL_BIT"));
-  if (clear_bit(&v, EGL_OPENGL_ES3_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_OPENGL_ES3_BIT"));
-  if (v != 0)
-    g_ptr_array_add(strings, egl_hexadecimal_to_string(v));
-  g_ptr_array_add(strings, nullptr);
-
-  return g_strjoinv("|", reinterpret_cast<gchar**>(strings->pdata));
-}
-
-// Converts an EGL surface type bitfield to a string.
-static gchar* egl_surface_type_to_string(EGLint value) {
-  EGLint v = value;
-  g_autoptr(GPtrArray) strings = g_ptr_array_new_with_free_func(g_free);
-  if (clear_bit(&v, EGL_PBUFFER_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_PBUFFER_BIT"));
-  if (clear_bit(&v, EGL_PIXMAP_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_PIXMAP_BIT"));
-  if (clear_bit(&v, EGL_WINDOW_BIT))
-    g_ptr_array_add(strings, g_strdup("EGL_WINDOW_BIT"));
-  if (v != 0)
-    g_ptr_array_add(strings, egl_hexadecimal_to_string(v));
-  g_ptr_array_add(strings, nullptr);
-
-  return g_strjoinv("|", reinterpret_cast<gchar**>(strings->pdata));
-}
-
-// Generates a string containing information about the given EGL configuration.
-static gchar* egl_config_to_string(EGLDisplay display, EGLConfig config) {
-  struct {
-    EGLint attribute;
-    const gchar* name;
-    gchar* (*to_string)(EGLint value);
-  } config_items[] = {{
-                          EGL_CONFIG_ID,
-                          "EGL_CONFIG_ID",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_BUFFER_SIZE,
-                          "EGL_BUFFER_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_COLOR_BUFFER_TYPE,
-                          "EGL_COLOR_BUFFER_TYPE",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_TRANSPARENT_TYPE,
-                          "EGL_TRANSPARENT_TYPE",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_LEVEL,
-                          "EGL_LEVEL",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_RED_SIZE,
-                          "EGL_RED_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_GREEN_SIZE,
-                          "EGL_GREEN_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_BLUE_SIZE,
-                          "EGL_BLUE_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_ALPHA_SIZE,
-                          "EGL_ALPHA_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_DEPTH_SIZE,
-                          "EGL_DEPTH_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_STENCIL_SIZE,
-                          "EGL_STENCIL_SIZE",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_SAMPLES,
-                          "EGL_SAMPLES",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_SAMPLE_BUFFERS,
-                          "EGL_SAMPLE_BUFFERS",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_NATIVE_VISUAL_ID,
-                          "EGL_NATIVE_VISUAL_ID",
-                          egl_hexadecimal_to_string,
-                      },
-                      {
-                          EGL_NATIVE_VISUAL_TYPE,
-                          "EGL_NATIVE_VISUAL_TYPE",
-                          egl_hexadecimal_to_string,
-                      },
-                      {
-                          EGL_NATIVE_RENDERABLE,
-                          "EGL_NATIVE_RENDERABLE",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_CONFIG_CAVEAT,
-                          "EGL_CONFIG_CAVEAT",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_BIND_TO_TEXTURE_RGB,
-                          "EGL_BIND_TO_TEXTURE_RGB",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_BIND_TO_TEXTURE_RGBA,
-                          "EGL_BIND_TO_TEXTURE_RGBA",
-                          egl_enum_to_string,
-                      },
-                      {
-                          EGL_RENDERABLE_TYPE,
-                          "EGL_RENDERABLE_TYPE",
-                          egl_renderable_type_to_string,
-                      },
-                      {
-                          EGL_CONFORMANT,
-                          "EGL_CONFORMANT",
-                          egl_renderable_type_to_string,
-                      },
-                      {
-                          EGL_SURFACE_TYPE,
-                          "EGL_SURFACE_TYPE",
-                          egl_surface_type_to_string,
-                      },
-                      {
-                          EGL_MAX_PBUFFER_WIDTH,
-                          "EGL_MAX_PBUFFER_WIDTH",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_MAX_PBUFFER_HEIGHT,
-                          "EGL_MAX_PBUFFER_HEIGHT",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_MAX_PBUFFER_PIXELS,
-                          "EGL_MAX_PBUFFER_PIXELS",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_MIN_SWAP_INTERVAL,
-                          "EGL_MIN_SWAP_INTERVAL",
-                          egl_decimal_to_string,
-                      },
-                      {
-                          EGL_MAX_SWAP_INTERVAL,
-                          "EGL_MAX_SWAP_INTERVAL",
-                          egl_decimal_to_string,
-                      },
-                      {EGL_NONE, nullptr, nullptr}};
-
-  g_autoptr(GPtrArray) strings = g_ptr_array_new_with_free_func(g_free);
-  for (int i = 0; config_items[i].attribute != EGL_NONE; i++) {
-    EGLint value;
-    if (!eglGetConfigAttrib(display, config, config_items[i].attribute, &value))
-      continue;
-    g_autofree gchar* value_string = config_items[i].to_string(value);
-    if (value_string == nullptr)
-      value_string = egl_hexadecimal_to_string(value);
-    g_ptr_array_add(
-        strings, g_strdup_printf("%s=%s", config_items[i].name, value_string));
-  }
-  g_ptr_array_add(strings, nullptr);
-
-  return g_strjoinv(" ", reinterpret_cast<gchar**>(strings->pdata));
-}
-
 // Creates a resource surface.
 static void create_resource_surface(FlRenderer* self, EGLConfig config) {
   FlRendererPrivate* priv =
@@ -308,14 +31,14 @@ static void create_resource_surface(FlRenderer* self, EGLConfig config) {
   priv->resource_surface = eglCreatePbufferSurface(priv->egl_display, config,
                                                    resource_context_attribs);
   if (priv->resource_surface == nullptr) {
-    g_warning("Failed to create EGL resource surface: %s", get_egl_error());
+    g_warning("Failed to create EGL resource surface: %s", egl_error_to_string(eglGetError()));
     return;
   }
 
   priv->resource_context = eglCreateContext(
       priv->egl_display, config, priv->egl_context, context_attributes);
   if (priv->resource_context == nullptr)
-    g_warning("Failed to create EGL resource context: %s", get_egl_error());
+    g_warning("Failed to create EGL resource context: %s", egl_error_to_string(eglGetError()));
 }
 
 // Default implementation for the start virtual method.
@@ -353,12 +76,12 @@ static gboolean fl_renderer_real_start(FlRenderer* self, GError** error) {
   if (!eglChooseConfig(priv->egl_display, attributes, &egl_config, 1,
                        &n_config)) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to choose EGL config: %s", get_egl_error());
+                "Failed to choose EGL config: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
   if (n_config == 0) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to find appropriate EGL config: %s", get_egl_error());
+                "Failed to find appropriate EGL config: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
   if (!eglBindAPI(EGL_OPENGL_ES_API)) {
@@ -366,7 +89,7 @@ static gboolean fl_renderer_real_start(FlRenderer* self, GError** error) {
         egl_config_to_string(priv->egl_display, egl_config);
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
                 "Failed to bind EGL OpenGL ES API using configuration (%s): %s",
-                config_string, get_egl_error());
+                config_string, egl_error_to_string(eglGetError()));
     return FALSE;
   }
 
@@ -377,7 +100,7 @@ static gboolean fl_renderer_real_start(FlRenderer* self, GError** error) {
         egl_config_to_string(priv->egl_display, egl_config);
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
                 "Failed to create EGL surface using configuration (%s): %s",
-                config_string, get_egl_error());
+                config_string, egl_error_to_string(eglGetError()));
     return FALSE;
   }
   EGLint context_attributes[] = {EGL_CONTEXT_CLIENT_VERSION, 2, EGL_NONE};
@@ -388,7 +111,7 @@ static gboolean fl_renderer_real_start(FlRenderer* self, GError** error) {
         egl_config_to_string(priv->egl_display, egl_config);
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
                 "Failed to create EGL context using configuration (%s): %s",
-                config_string, get_egl_error());
+                config_string, egl_error_to_string(eglGetError()));
     return FALSE;
   }
 
@@ -422,7 +145,7 @@ gboolean fl_renderer_make_current(FlRenderer* self, GError** error) {
   if (!eglMakeCurrent(priv->egl_display, priv->egl_surface, priv->egl_surface,
                       priv->egl_context)) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to make EGL context current: %s", get_egl_error());
+                "Failed to make EGL context current: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
 
@@ -439,7 +162,7 @@ gboolean fl_renderer_make_resource_current(FlRenderer* self, GError** error) {
   if (!eglMakeCurrent(priv->egl_display, priv->resource_surface,
                       priv->resource_surface, priv->resource_context)) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to make EGL context current: %s", get_egl_error());
+                "Failed to make EGL context current: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
 
@@ -453,7 +176,7 @@ gboolean fl_renderer_clear_current(FlRenderer* self, GError** error) {
   if (!eglMakeCurrent(priv->egl_display, EGL_NO_SURFACE, EGL_NO_SURFACE,
                       EGL_NO_CONTEXT)) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to clear EGL context: %s", get_egl_error());
+                "Failed to clear EGL context: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
 
@@ -471,7 +194,7 @@ gboolean fl_renderer_present(FlRenderer* self, GError** error) {
 
   if (!eglSwapBuffers(priv->egl_display, priv->egl_surface)) {
     g_set_error(error, fl_renderer_error_quark(), FL_RENDERER_ERROR_FAILED,
-                "Failed to swap EGL buffers: %s", get_egl_error());
+                "Failed to swap EGL buffers: %s", egl_error_to_string(eglGetError()));
     return FALSE;
   }
